Добавил команду download, делал рефакторинг: создал класс FileWorker для обработки отправки и получения файла.
